fix: signal-slot connect failure
It was really a bad experience. My Qt seems to be broken after a build,
and both QTimer::timeout and QAbstractButton::clicked, or pressed() and
so on, is reported to be not seen in the given object by the connect()
method. That would be absolutely crazy! And from then on, the error
always exists even when I rollback to an old but working version.

I asked my friend and he also failed to run on the latest version, but
the old versions behaves well on his machine. I have no ideas on how to
deals with the case but use the old form for the QObject::connect(),
which will search for the signal & slot in the runtime.

It is necessary to state that the problem must be carefully traded and
come up wil a pretty solution in the future. After all, connect by a
pair of strings generated by macros indeed cost extra time and energy.

NOTE: relative question can be found at 61879664 in the stackoverflow